requirements.txt
mcp>=1.2.0
feedparser>=6.0.11
pydantic>=2.7
pyyaml>=6.0
requests>=2.32
python-dateutil>=2.9.0
beautifulsoup4>=4.12
python-dotenv>=1.0.0
